+++ R/pattern-array-gradient.R
@@ -6,6 +6,9 @@
 #' @param fill2 Second colour
 #' @param orientation vertical, horizontal, or radial
 #' @param aspect_ratio Override aspect ratio
+#' @param use_R4.1_gradients Whether to use the gradient feature introduced in R v4.1
+#'                           or use a `rasterGrob` approximation.
+#'                           Note not all graphic devices support the grid gradient feature.
 #' @param key_scale_factor Additional scale factor for legend
 #' @param res Assumed resolution (in pixels per graphic device inch) to use when creating array pattern.
 #' @return A grid grob object invisibly.  If `draw` is `TRUE` then also draws to the graphic device as a side effect.
@@ -20,11 +23,14 @@
 grid.pattern_gradient <- function(x = c(0, 0, 1, 1), y = c(1, 0, 0, 1), id = 1L, ...,
                                   fill = gp$fill %||% "grey80", fill2 = "#4169E1",
                                   orientation = "vertical", alpha = gp$alpha %||% NA_real_,
+                                  use_R4.1_gradients = getOption("ggpattern_use_R4.1_gradients",
+                                                                 getOption("ggpattern_use_R4.1_features")),
                                   aspect_ratio = 1, key_scale_factor = 1, res = 72,
                                   default.units = "npc", name = NULL, gp = gpar(), draw = TRUE, vp = NULL) {
     grid.pattern("gradient", x, y, id,
                  fill = fill, fill2 = fill2,
                  orientation = orientation, alpha = alpha,
+                 use_R4.1_gradients = use_R4.1_gradients,
                  aspect_ratio = aspect_ratio, key_scale_factor = key_scale_factor, res = res,
                  default.units = default.units, name = name, gp = gp , draw = draw, vp = vp)
 }
@@ -66,6 +72,48 @@ create_gradient_img <- function(width       = 100,
   img
 }
 
+create_pattern_gradient <- function(params, boundary_df, aspect_ratio, legend = FALSE) {
+    if (params$pattern_use_R4.1_gradients) {
+        create_gradient_as_geometry(params, boundary_df, aspect_ratio, legend)
+    } else {
+        create_pattern_array(params, boundary_df, aspect_ratio, legend, create_gradient_as_array)
+    }
+}
+
+create_gradient_as_geometry <- function(params, boundary_df, aspect_ratio, legend) {
+  orientation <- check_default(params$pattern_orientation,
+                               options = c('vertical', 'horizontal', 'radial'))
+  colour1     <- params$pattern_fill
+  colour2     <- params$pattern_fill2
+
+  x_min <- min(boundary_df$x)
+  x_max <- max(boundary_df$x)
+  x_med <- 0.5 * (x_min + x_max)
+  y_min <- min(boundary_df$y)
+  y_max <- max(boundary_df$y)
+  y_med <- 0.5 * (y_min + y_max)
+  x_range <- convertX(unit(x_max - x_min, "npc"), "in")
+  y_range <- convertY(unit(y_max - y_min, "npc"), "in")
+  gradient <- switch(orientation,
+     horizontal = linearGradient(c(colour1, colour2),
+                                 x1 = x_min, y1 = y_med,
+                                 x2 = x_max, y2 = y_med),
+     radial = radialGradient(c(colour1, colour2),
+                             cx1 = x_med, cy1 = y_med,
+                             cx2 = x_med, cy2 = y_med,
+                             r2 = 0.5 * max(x_range, y_range),
+                             extend = "none"
+                             ),
+     vertical = linearGradient(c(colour1, colour2),
+                               x1 = x_med, y1 = y_min,
+                               x2 = x_med, y2 = y_max)
+  )
+
+  gp <- gpar(col = NA, fill = gradient)
+
+  convert_polygon_df_to_polygon_grob(boundary_df, gp = gp)
+}
+
 #' A shim to go between the main pattern function for an image, and the
 #' specific pattern functon for an image.
 #'
@@ -81,7 +129,8 @@ create_gradient_img <- function(width       = 100,
 #' @noRd
 create_gradient_as_array <- function(width, height, params, legend) {
 
-  orientation <- check_default(params$pattern_orientation, options = c('vertical', 'horizontal', 'radial'))
+  orientation <- check_default(params$pattern_orientation,
+                               options = c('vertical', 'horizontal', 'radial'))
   colour1     <- params$pattern_fill
   colour2     <- params$pattern_fill2